requirements.txt
django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
mysqlclient
python-decouple
dj-database-url
gunicorn
Pillow
